track shot lifetimes on the server and end them when it is time. only end shots that end on the server.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/branches/v2_99_shot_branch@17088 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/GameKeeper.cxx
@@ -570,7 +570,7 @@ bool GameKeeper::Player::addShot( int globalID, int localID, double startTime )
     lifeTime *= BZDB.eval(StateDatabase::BZDB_THIEFADLIFE);
 
   // fill up the slot
-  shotSlots[freeSlot].expireTime  = startTime + lifeTime;
+  shotSlots[freeSlot].expireTime  = (float)startTime + lifeTime;
   shotSlots[freeSlot].present     = true;
   shotSlots[freeSlot].running     = true;
